Refactor storage_defs out of runtime_types.
