+++ src/include/execution/sql/operators/hash_operators.h
@@ -2,6 +2,7 @@
 
 #include "common/hash_util.h"
 #include "execution/sql/runtime_types.h"
+#include "storage/storage_defs.h"
 
 namespace terrier::execution::sql {
