+++ src/include/execution/sql/runtime_types.h
@@ -1,11 +1,12 @@
 #pragma once
 
 #include <algorithm>
+#include <cstring>
 #include <string>
 #include <utility>
 
-#include "execution/util/string_heap.h"
-#include "storage/storage_defs.h"
+#include "common/strong_typedef.h"
+#include "execution/util/execution_common.h"
 
 namespace terrier::execution::sql {
 
@@ -73,12 +74,12 @@ class EXPORT Date {
    * @param seed The value to seed the hash with.
    * @return The hash value for this date instance.
    */
-  hash_t Hash(const hash_t seed) const { return common::HashUtil::HashCrc(value_, seed); }
+  hash_t Hash(const hash_t seed) const;
 
   /**
    * @return The hash value of this date instance.
    */
-  hash_t Hash() const { return Hash(0); }
+  hash_t Hash() const;
 
   /**
    * @return True if this date equals @em that date; false otherwise.
@@ -262,12 +263,12 @@ class EXPORT Timestamp {
    * @param seed The value to seed the hash with.
    * @return The hash value for this timestamp instance.
    */
-  hash_t Hash(const hash_t seed) const { return common::HashUtil::HashCrc(value_, seed); }
+  hash_t Hash(const hash_t seed) const;
 
   /**
    * @return The hash value of this timestamp instance.
    */
-  hash_t Hash() const { return Hash(0); }
+  hash_t Hash() const;
 
   /**
    * @return True if this timestamp equals @em that timestamp; false otherwise.
@@ -431,12 +432,12 @@ class EXPORT Decimal {
    * @param seed The value to seed the hash with.
    * @return The hash value for this decimal instance.
    */
-  hash_t Hash(const hash_t seed) const { return common::HashUtil::HashCrc(value_); }
+  hash_t Hash(const hash_t seed) const;
 
   /**
    * @return The hash value of this decimal instance.
    */
-  hash_t Hash() const { return Hash(0); }
+  hash_t Hash() const;
 
   /**
    * Add the encoded decimal value @em that to this decimal value.
@@ -496,61 +497,6 @@ class EXPORT Decimal {
 using Decimal32 = Decimal<int32_t>;
 using Decimal64 = Decimal<int64_t>;
 using Decimal128 = Decimal<int128_t>;
-
-//===----------------------------------------------------------------------===//
-//
-// Variable-length values
-//
-//===----------------------------------------------------------------------===//
-
-/**
- * A container for varlens.
- */
-class EXPORT VarlenHeap {
- public:
-  /**
-   * Allocate memory from the heap whose contents will be filled in by the user BEFORE creating a varlen.
-   * @param len The length of the varlen to allocate.
-   * @return The character byte array.
-   */
-  char *PreAllocate(std::size_t len) { return heap_.Allocate(len); }
-
-  /**
-   * Allocate a varlen from this heap whose contents are the same as the input string.
-   * @param str The string to copy into the heap.
-   * @param len The length of the input string.
-   * @return A varlen.
-   */
-  storage::VarlenEntry AddVarlen(const char *str, std::size_t len) {
-    auto *content = heap_.AddString(std::string_view(str, len));
-    return storage::VarlenEntry::Create(reinterpret_cast<byte *>(content), len, false);
-  }
-
-  /**
-   * Allocate and return a varlen from this heap whose contents as the same as the input string.
-   * @param string The string to copy into the heap.
-   * @return A varlen.
-   */
-  storage::VarlenEntry AddVarlen(const std::string &string) { return AddVarlen(string.c_str(), string.length()); }
-
-  /**
-   * Add a copy of the given varlen into this heap.
-   * @param other The varlen to copy into this heap.
-   * @return A new varlen entry.
-   */
-  storage::VarlenEntry AddVarlen(const storage::VarlenEntry &other) {
-    return AddVarlen(reinterpret_cast<const char *>(other.Content()), other.Size());
-  }
-
-  /**
-   * Destroy all heap-allocated varlens.
-   */
-  void Destroy() { heap_.Destroy(); }
-
- private:
-  // Internal heap of strings
-  util::StringHeap heap_;
-};
 
 /**
  * Simple structure representing a blob.
