+++ test/execution/sql_varlen_entry_test.cpp
@@ -5,6 +5,7 @@
 
 #include "execution/sql/runtime_types.h"
 #include "execution/tpl_test.h"
+#include "storage/storage_defs.h"
 
 namespace terrier::execution::sql::test {
